Updating daily BSM full import
It was previously being run everyday except Thursday, but since it runs at 3 am that didn’t make sense. We were trying to avoid running it while the full import was running, which starts Thursday night at 8pm. This fixes the cron jobs to run every day except Friday, which is when the full import will still be running.